requirements.txt
matplotlib==3.0.3
torch==1.2.0
torchvision==0.4.0
opencv_contrib_python==4.1.0.25
tqdm==4.32.1
numpy==1.16.2
pandas==0.24.2
scikit_image==0.14.2
blessings==1.7
imageio==2.5.0
joblib==0.13.2
path.py==12.4.0
Pillow==6.2.1
progressbar33==2.4
scikit_learn==0.22
tensorboardX==1.9
